node.js server up and running
get and post routes now work. When a user enters information and submits, data is requested from an API, then that data is sent to the server and combined with the user's data. The full data object is sent back to the client and new elements are displayed on the web page.